Refactor blog post loading logic into separate lib
